Data grab overhaul
Both methods now grab the most recent trading day along with the previous trading day and stores it in a list for accurate information

Previously the program would grab the most recent trading day's (or previous depending on weekend or not) information and then reading that along with whatever trading day was last downloaded in the file instead of *specifically* the previous trading day. Having the program remain that way would've made it so you would have to run the program every day to create relevant and accurate information. 

exampleTwo requires lines upon lines of hardcoded exceptions if I were to include holidays when the stock market is closed. exampleOne is not subject to this.